Parallelize multi-bit B2A converter (#24)
Summary:
Pull Request resolved: https://github.com/facebookresearch/CrypTen/pull/24

Pull Request resolved: https://github.com/fairinternal/CrypTen/pull/155

Vectorized B2A calls in B2A converter in order to parallelize rounds. We skipped this optimization in earlier versions because it was not relevant to our code runs.

I did not change or add to the tests here because our tests cover bits = 1 and bits = 64. All other values for bits do not have nice analogs to check against floating point values. I think this is fine.

Reviewed By: shubho

Differential Revision: D18246140

fbshipit-source-id: 801af3861170306805705f3c6c92ece272fe565f