Remove unnecessary divides in comparrisons
Summary:
Functions like ReLU were computing [x] * [x > 0] using a scaled [x] and a scaled [x > 0]. However we can avoid downscaling the result of these functions by removing the upscaling of [x > 0].

(i.e. before this we were multiplying and dividing by the encoder scale - but this is unnecessary and increases the probability of wraparound errors).

Reviewed By: shubho

Differential Revision: D18818113

fbshipit-source-id: 70776d1f791e8cc7c02aa35693f6cca6a74c44c1